+++ cezo_fl/coordinate_gradient_estimator_test.py
@@ -3,10 +3,10 @@ import torch
 from torch import nn
 from torch.optim import SGD
 
-from gradient_estimators.coordinate_gradient_estimator import (
-    get_parameter_indices_for_ith_elem,
-    CoordinateGradientEstimator as CGE,
-)
+from cezo_fl.coordinate_gradient_estimator import \
+    CoordinateGradientEstimator as CGE
+from cezo_fl.coordinate_gradient_estimator import \
+    get_parameter_indices_for_ith_elem
 
 
 class LinearReg(nn.Module):
