+++ cezo_fl/models/cnn_cifar10.py
