+++ cezo_fl/models/cnn_mnist.py
@@ -1,6 +1,5 @@
 from torch import nn
 
-
 # This commented CNN have exploding gradient issue when batch_size = 4
 # class CNN_MNIST(nn.Module):
 #     def __init__(self):
