+++ cezo_fl/run_client_jobs_test.py
@@ -1,18 +1,20 @@
 from unittest.mock import MagicMock
+
+import numpy as np
 import pytest
 import torch
 import torchvision
 import torchvision.transforms as transforms
-import numpy as np
+from torch.optim import SGD
 
 from cezo_fl.client import ResetClient
-from cezo_fl.run_client_jobs import parallalizable_client_job, execute_sampled_clients
-from models.cnn_mnist import CNN_MNIST
+from cezo_fl.models.cnn_mnist import CNN_MNIST
+from cezo_fl.random_gradient_estimator import RandomGradientEstimator as RGE
+from cezo_fl.run_client_jobs import (execute_sampled_clients,
+                                     parallalizable_client_job)
+from cezo_fl.util.metrics import accuracy
 from config import FakeArgs
 from preprocess import preprocess
-from gradient_estimators.random_gradient_estimator import RandomGradientEstimator as RGE
-from torch.optim import SGD
-from shared.metrics import accuracy
 
 
 def get_mnist_data_loader():
