+++ cezo_fl/util/checkpoint.py
@@ -1,5 +1,6 @@
-import torch
 import os
+
+import torch
 
 from config import get_args_dict
